split out the secret model from the db and table services. This is to support a better more modular structure and more tables such as a table of unique tags, unique vaults, and device sync rules.
